Move activities route to kapture-backend
