+++ ui/src/implementations/kapture-backend/routes/activities/index.js
